Added another activity(MainPasswordHub) an image adapter for that activity as well as its layout.
The adapter seems to be working fine but we will need to add the images ourselves in the drawable directory, as to adding the pictures in the code, there's an example in the adapter.

This activity will be the main screen after the log in, the app will redirect here.